Feat: Add layers in Image Block in Gutenberg
